Rename s_module.rb file to scrape_init.rb
