+++ lib/scrape_init.rb
